+++ tests/regenerate_tests.py
@@ -5,6 +5,7 @@
 # SPDX-License-Identifier: MIT
 
 import re
+import sys
 from utils import *
 from pathlib import Path
 try:
@@ -38,7 +39,7 @@ def _cfg(config, path, default=None):
 def regenerate_expected_outputs():
 	test_root = Path(Path(__file__).parent).resolve()
 
-	for subdir in enumerate_directories(test_root, filter=lambda p: Path(p, r'poxy.toml').is_file()):
+	for subdir in enumerate_directories(test_root, filter=lambda p: p.name.startswith(r'test_')):
 
 		html_dir = Path(subdir, r'html')
 		xml_dir = Path(subdir, r'xml')
@@ -58,23 +59,25 @@ def regenerate_expected_outputs():
 			r'--noassets',  #
 			rf'--{"no-" if not output_html else ""}html',
 			rf'--{"no-" if not output_xml else ""}xml',
+			*sys.argv[1:]
 		]
 		if bool(_cfg(config, (r'xml', r'v2'), False)):
-			args.append(r'--experimental-xml-v2')
+			args.append(r'--xml-v2')
 
 		# delete all previous outputs
+		print(r"Deleting previous artefacts...")
 		delete_directory(html_dir, logger=True)
 		delete_directory(xml_dir, logger=True)
 		delete_directory(expected_html_dir, logger=True)
 		delete_directory(expected_xml_dir, logger=True)
 
 		# run poxy
-		print(rf"Regenerating {subdir}...")
+		print(rf"Running Poxy...")
 		run_poxy(subdir, *args)
 
 		# delete garbage
 		garbage = (
-			r'*.xslt', r'*.xsd', r'favicon*', r'search-v2.js',
+			r'*.xslt', r'*.xsd', r'favicon*', r'search-v2.js', r'Doxyfile*',
 			*(coerce_collection([r'garbage']) if r'garbage' in config else [])
 		)
 		garbage = (
@@ -84,35 +87,38 @@ def regenerate_expected_outputs():
 		for file in garbage:
 			delete_file(file, logger=True)
 
-		# process html files
+		# sanitize + fix sources of nondeterminism (so the tests work everywhere)
+		print(r"Sanitizing outputs...")
+		files = (
+			*(enumerate_files(html_dir, any=(r'*.html', r'*.xml')) if output_html else []),
+			*(enumerate_files(xml_dir, any=(r'*.html', r'*.xml')) if output_xml else []),
+		)
+		for path in files:
+			text = read_all_text_from_file(path)
+			text = text.replace(r'href="poxy/poxy.css"', r'href="../../../poxy/data/css/poxy.css"')
+			text = text.replace(r'src="poxy/poxy.js"', r'src="../../../poxy/data/poxy.js"')
+			text = text.replace(r'src="search-v2.js"', r'src="../../../poxy/data/m.css/documentation/search.js"')
+			text = re.sub(r'Poxy v[0-9]+[.][0-9]+[.][0-9]+', r'Poxy v0.0.0', text)
+			text = re.sub(r'version="\s*[0-9]+[.][0-9]+[.][0-9]+\s*"', r'version="0.0.0"', text)
+			text = re.sub(r'gitid="\s*[0-9a-fA-F]+\s*"', r'gitid="000000000000000000000000000000000000000"', text)
+			print(rf"Writing {path}")
+			with open(path, r'w', encoding=r'utf-8', newline='\n') as f:
+				f.write(text)
+
+		# move html files
 		if not output_html:
 			delete_directory(html_dir, logger=True)
 		else:
-			for path in enumerate_files(html_dir, any=(r'*.html')):
-				text = read_all_text_from_file(path)
-				text = text.replace(r'href="poxy/poxy.css"', r'href="../../../poxy/data/css/poxy.css"')
-				text = text.replace(r'src="poxy/poxy.js"', r'src="../../../poxy/data/poxy.js"')
-				text = text.replace(r'src="search-v2.js"', r'src="../../../poxy/data/m.css/documentation/search.js"')
-				text = re.sub(r'Poxy v[0-9]+[.][0-9]+[.][0-9]+', r'Poxy v0.0.0', text)
-				print(rf"Writing {path}")
-				with open(path, r'w', encoding=r'utf-8', newline='\n') as f:
-					f.write(text)
 			html_dir.rename(expected_html_dir)
 
-		# process xml files
+		# move xml files
 		if not output_xml:
 			delete_directory(xml_dir, logger=True)
 		else:
-			for path in enumerate_files(xml_dir, any=(r'*.xml')):
-				text = read_all_text_from_file(path)
-				text = re.sub(r'version="\s*[0-9]+[.][0-9]+[.][0-9]+\s*"', r'version="0.0.0"', text)
-				print(rf"Writing {path}")
-				with open(path, r'w', encoding=r'utf-8', newline='\n') as f:
-					f.write(text)
 			xml_dir.rename(expected_xml_dir)
 
 
 
-if __name__ == '__main__':
+if __name__ == r'__main__':
 	with ScopeTimer(r'Regenerating test outputs'):
 		regenerate_expected_outputs()